speculation-rules/prefetch/prefetch-status.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js?feature=bidi"></script>
<script src="/resources/testdriver-extra.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.sub.js"></script>


<script>
  setup(() => assertSpeculationRulesIsSupported());
  
  promise_test(async t => {
    await test_driver.bidi.speculation.prefetch_status_updated.subscribe();
    
    let Got_statuses;
    const Got_statuses_promise = new Promise((resolve) => {
      Got_statuses = { resolve };
    });
    
    const eventStatus = [];

    let agent = await spawnWindow(t, { protocol: "https" });
    let nextUrl = agent.getExecutorURL({ protocol: "https", page: 2 });

    const removeHandler = test_driver.bidi.speculation.prefetch_status_updated.on(async (event) => {
      console.log('I GOT AN EVENT', event);
      eventStatus.push(event.status); // Push just the status, not the whole event
      
      if (event.status === "running") {
        console.log("Prefetch is running");
      } else if (event.status === "ready") {
        console.log("Prefetch is ready");
        await agent.navigate(nextUrl);
      } else if (event.status === "success") {
        console.log("Prefetch succeeded");
        assert_equals(eventStatus.length, 3, "Should have received 3 status events");
        assert_array_equals(eventStatus, ["running", "ready", "success"], "Should receive statuses in correct order");
        
        const headers = await agent.getRequestHeaders();
        assert_prefetched(headers, "Prefetch should work for HTTPS urls.");
        
        removeHandler();
        Got_statuses.resolve();
      }
    });
    
    await agent.forceSinglePrefetch(nextUrl, {}, true);
    await Got_statuses_promise;
  }, "Prefetch status updates should be received");
</script>
</html>
